+++ e2e/forwarding_test.go
@@ -31,7 +31,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/gogoproto/jsonpb"
 	transfertypes "github.com/cosmos/ibc-go/v8/modules/apps/transfer/types"
-	"github.com/icza/dyno"
 	forwardingtypes "github.com/noble-assets/forwarding/v2/types"
 	"github.com/strangelove-ventures/interchaintest/v8"
 	"github.com/strangelove-ventures/interchaintest/v8/chain/cosmos"
@@ -46,7 +45,7 @@ import (
 func TestRegisterOnNoble(t *testing.T) {
 	t.Parallel()
 
-	ctx, noble, gaia, _, _, sender, _, receiver := ForwardingSuite(t, nil)
+	ctx, noble, gaia, _, _, sender, _, receiver := ForwardingSuite(t, []string{"uusdc"})
 	validator := noble.Validators[0]
 
 	address, exists := ForwardingAccount(t, ctx, validator, receiver, "")
@@ -89,7 +88,7 @@ func TestRegisterOnNoble(t *testing.T) {
 func TestRegisterOnNobleSignerlessly(t *testing.T) {
 	t.Parallel()
 
-	ctx, noble, _, _, _, sender, _, receiver := ForwardingSuite(t, nil)
+	ctx, noble, _, _, _, sender, _, receiver := ForwardingSuite(t, []string{"uusdc"})
 	validator := noble.Validators[0]
 
 	address, exists := ForwardingAccount(t, ctx, validator, receiver, "")
@@ -117,7 +116,9 @@ func TestRegisterOnNobleSignerlessly(t *testing.T) {
 func TestRegisterViaTransfer(t *testing.T) {
 	t.Parallel()
 
-	ctx, noble, gaia, _, _, _, _, receiver := ForwardingSuite(t, nil)
+	uatom := transfertypes.DenomTrace{Path: "transfer/channel-0", BaseDenom: "uatom"}.IBCDenom()
+
+	ctx, noble, gaia, _, _, _, _, receiver := ForwardingSuite(t, []string{uatom})
 	validator := noble.Validators[0]
 
 	address, exists := ForwardingAccount(t, ctx, validator, receiver, "")
@@ -149,10 +150,7 @@ func TestRegisterViaTransfer(t *testing.T) {
 	stats := ForwardingStats(t, ctx, validator)
 	require.Equal(t, uint64(1), stats.NumOfAccounts)
 	require.Equal(t, uint64(1), stats.NumOfForwards)
-	require.Equal(t, sdk.NewCoins(sdk.NewCoin(transfertypes.DenomTrace{
-		Path:      "transfer/channel-0",
-		BaseDenom: "uatom",
-	}.IBCDenom(), math.NewInt(100_000))), stats.TotalForwarded)
+	require.Equal(t, sdk.NewCoins(sdk.NewCoin(uatom, math.NewInt(100_000))), stats.TotalForwarded)
 }
 
 func TestRegisterViaPacket(t *testing.T) {
@@ -162,7 +160,7 @@ func TestRegisterViaPacket(t *testing.T) {
 func TestFrontRunAccount(t *testing.T) {
 	t.Parallel()
 
-	ctx, noble, gaia, _, _, sender, _, receiver := ForwardingSuite(t, nil)
+	ctx, noble, gaia, _, _, sender, _, receiver := ForwardingSuite(t, []string{"uusdc"})
 	validator := noble.Validators[0]
 
 	address, exists := ForwardingAccount(t, ctx, validator, receiver, "")
@@ -209,7 +207,7 @@ func TestFrontRunAccount(t *testing.T) {
 func TestClearAccount(t *testing.T) {
 	t.Parallel()
 
-	ctx, noble, gaia, rly, execReporter, sender, _, receiver := ForwardingSuite(t, nil)
+	ctx, noble, gaia, rly, execReporter, sender, _, receiver := ForwardingSuite(t, []string{"uusdc"})
 	validator := noble.Validators[0]
 
 	require.NoError(t, rly.StopRelayer(ctx, execReporter))
@@ -277,7 +275,7 @@ func TestClearAccount(t *testing.T) {
 func TestFallbackAccount(t *testing.T) {
 	t.Parallel()
 
-	ctx, noble, gaia, rly, execReporter, sender, fallback, receiver := ForwardingSuite(t, nil)
+	ctx, noble, gaia, rly, execReporter, sender, fallback, receiver := ForwardingSuite(t, []string{"uusdc"})
 	validator := noble.Validators[0]
 
 	require.NoError(t, rly.StopRelayer(ctx, execReporter))
@@ -353,7 +351,7 @@ func TestFallbackAccount(t *testing.T) {
 func TestAllowedDenoms(t *testing.T) {
 	t.Parallel()
 
-	ctx, noble, gaia, _, _, sender, fallback, receiver := ForwardingSuite(t, &[]string{"uusdc"})
+	ctx, noble, gaia, _, _, sender, fallback, receiver := ForwardingSuite(t, []string{"uusdc"})
 	validator := noble.Validators[0]
 
 	res := ForwardingDenoms(t, ctx, validator)
@@ -392,18 +390,6 @@ func TestAllowedDenoms(t *testing.T) {
 	balance, err = noble.BankQueryAllBalances(ctx, address)
 	require.NoError(t, err)
 	require.Equal(t, sdk.NewCoins(sdk.NewCoin(uatom, math.NewInt(100_000))), balance)
-
-	_, err = validator.ExecTx(ctx, sender.KeyName(), "forwarding", "clear-account", address, "--fallback")
-	require.NoError(t, err)
-	require.NoError(t, testutil.WaitForBlocks(ctx, 10, noble, gaia))
-
-	balance, err = noble.BankQueryAllBalances(ctx, address)
-	require.NoError(t, err)
-	require.True(t, balance.IsZero())
-
-	fallbackBalance, err := noble.BankQueryAllBalances(ctx, fallback.FormattedAddress())
-	require.NoError(t, err)
-	require.Equal(t, sdk.NewCoins(sdk.NewCoin(uatom, math.NewInt(100_000))), fallbackBalance)
 
 	stats := ForwardingStats(t, ctx, validator)
 	require.Equal(t, uint64(1), stats.NumOfAccounts)
@@ -474,7 +460,7 @@ func TxFee(t *testing.T, ctx context.Context, validator *cosmos.ChainNode, hash 
 	return res.Tx.AuthInfo.Fee.Amount
 }
 
-func ForwardingSuite(t *testing.T, denoms *[]string) (ctx context.Context, noble *cosmos.CosmosChain, gaia *cosmos.CosmosChain, relayer *rly.CosmosRelayer, execReporter *testreporter.RelayerExecReporter, sender ibc.Wallet, fallback ibc.Wallet, receiver ibc.Wallet) {
+func ForwardingSuite(t *testing.T, denoms []string) (ctx context.Context, noble *cosmos.CosmosChain, gaia *cosmos.CosmosChain, relayer *rly.CosmosRelayer, execReporter *testreporter.RelayerExecReporter, sender ibc.Wallet, fallback ibc.Wallet, receiver ibc.Wallet) {
 	ctx = context.Background()
 	logger := zaptest.NewLogger(t)
 	reporter := testreporter.NewNopReporter()
@@ -508,25 +494,9 @@ func ForwardingSuite(t *testing.T, denoms *[]string) (ctx context.Context, noble
 				TrustingPeriod: "504h",
 				NoHostMount:    false,
 				ModifyGenesis: func(cfg ibc.ChainConfig, bz []byte) ([]byte, error) {
-					if denoms == nil {
-						return bz, nil
-					}
-
-					gen := make(map[string]interface{})
-					if err := json.Unmarshal(bz, &gen); err != nil {
-						return nil, fmt.Errorf("failed to unmarshal genesis: %w", err)
-					}
-
-					if err := dyno.Set(gen, denoms, "app_state", "forwarding", "allowed_denoms"); err != nil {
-						return nil, fmt.Errorf("failed to set forwarding allowed denoms in genesis: %w", err)
-					}
-
-					bz, err := json.Marshal(&gen)
-					if err != nil {
-						return nil, fmt.Errorf("failed to marshal genesis: %w", err)
-					}
-
-					return bz, nil
+					return cosmos.ModifyGenesis([]cosmos.GenesisKV{
+						cosmos.NewGenesisKV("app_state.forwarding.allowed_denoms", denoms),
+					})(cfg, bz)
 				},
 			},
 		},